Texturversuche für Treemap!
Timeline Querlinien

minTime als Filterzeit für letzen Tag // muss noch auf Tags angewendet
werden und ausschaltbar werden

buttons mit '9' ein/ausschaltbar

Treemap kann auch z-Achse haben

+++ TagExplorerProcessing2/src/tagexplorerprocessing2/SQLhelper.java
@@ -505,7 +505,7 @@ public class SQLhelper {
 
 				}
 				if (p5.showTimeline) {
-					file.z = -p5.timeline.mapExp(p5.getNewestDate(file));
+					file.z = (float) -p5.timeline.mapExp(p5.getNewestDate(file));
 				}
 			}
